Backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
boto3==1.34.0
Werkzeug==3.0.1
pytz==2024.1
gunicorn==21.2.0
